Interactive: Fix NPE on unsolicited stdout/stderr
If output is received while command == null (i.e. there is no outstanding
command), the app will crash.  One easy way to reproduce this is to
call Builder.open() and then deny the SU request through the GUI.  The su
UNIX command will print "Permission denied" to stderr and this generates
an exception.

So, let's just discard any output that is printed outside the context
of a valid command.