Implement fine-grained logging
Some applications send a large number of commands to the shell in rapid
succession, and the default behavior of logging all commands + output in
debug mode can significantly affect performance and clutter up the logcat
output.  Provide a way for the application to disable the noisiest logs
(commands and output) while retaining sanity checks and error messages.